api/requirements.txt
fastapi
uvicorn
requests
guardrails-ai
instructor
